Fix bug in button timer
Issues was that the duplication button is not supposed to be locked
because the final button after that button is supposed (and is) locked.
This worked, however in change the text back to Submit after the timer
has ended this exclusion was not present. This has now been fixed, so
the button will read the correct text: "Mark as duplicate" instead as
intended.